+++ frontend/src/routes/homePageF/assetF/Asset.jsx
